+++ moban/constants.py
@@ -64,6 +64,7 @@ LABEL_ACTION = "action"
 LABEL_SOURCE = "source"
 LABEL_DEST = "destination"
 LABEL_FORCE_TEMPLATE_TYPE = "force_template_type"
+LABEL_TEMPLATE_TYPES = "template_types"
 
 # error messages
 ERROR_DATA_FILE_NOT_FOUND = "Both %s and %s does not exist"
@@ -98,6 +99,10 @@ PYPI_REQUIRE = "PYPI"
 PYPI_PACKAGE_NAME = "name"
 REQUIRE_TYPE = "type"
 
+# Template types
+TEMPLATE_TYPES_BASE_TYPE = "base_type"
+TEMPLATE_TYPES_OPTIONS = "options"
+TEMPLATE_TYPES_FILE_EXTENSIONS = "file_extensions"
 
 # Extension
 JINJA_FILTER_EXTENSION = "jinja_filter"