+++ tests/test_template.py
@@ -1,14 +1,17 @@
 import os
 
 from mock import patch
+from nose.tools import eq_
 
 from moban.plugins import ENGINES
 from moban.definitions import TemplateTarget
+from moban.jinja2.engine import Engine
+from moban.data_loaders.yaml import open_yaml
 
 MODULE = "moban.plugins.template"
 
 
-@patch(MODULE + ".TemplateEngine._render_with_finding_data_first")
+@patch(MODULE + ".MobanEngine._render_with_finding_data_first")
 def test_do_templates_1(_do_templates_with_more_shared_data):
     jobs = [
         TemplateTarget("1.template", "data.yml", "1.output"),
@@ -31,7 +34,7 @@ def test_do_templates_1(_do_templates_with_more_shared_data):
     _do_templates_with_more_shared_data.assert_called_with(expected)
 
 
-@patch(MODULE + ".TemplateEngine._render_with_finding_template_first")
+@patch(MODULE + ".MobanEngine._render_with_finding_template_first")
 def test_do_templates_2(_do_templates_with_more_shared_templates):
     jobs = [
         TemplateTarget("1.template", "data1.yml", "1.output"),
@@ -80,3 +83,33 @@ def test_do_templates_with_more_shared_data():
         content = f.read()
         assert content == "hello world ox"
     os.unlink("test")
+
+
+def test_get_user_defined_engine():
+    test_fixture = os.path.join(
+        "tests", "fixtures", "mobanengine", "sample_template_type.yml"
+    )
+    template_types = open_yaml(test_fixture)
+    ENGINES.register_options(template_types["template_types"])
+    engine = ENGINES.get_engine("custom_jinja", ".", ".")
+    eq_(engine.engine.__class__, Engine)
+
+
+def test_custom_file_extension_is_assocated_with_user_defined_engine():
+    test_fixture = os.path.join(
+        "tests", "fixtures", "mobanengine", "sample_template_type.yml"
+    )
+    template_types = open_yaml(test_fixture)
+    ENGINES.register_options(template_types["template_types"])
+    template_type = ENGINES.get_primary_key("demo_file_suffix")
+    eq_("custom_jinja", template_type)
+
+
+def test_built_in_jinja2_file_extension_still_works():
+    test_fixture = os.path.join(
+        "tests", "fixtures", "mobanengine", "sample_template_type.yml"
+    )
+    template_types = open_yaml(test_fixture)
+    ENGINES.register_options(template_types["template_types"])
+    template_type = ENGINES.get_primary_key("jj2")
+    eq_("jinja2", template_type)