+++ CHANGELOG.rst
@@ -1,6 +1,19 @@
 Change log
 ================================================================================
 
+0.4.1 - 28.02.2019
+--------------------------------------------------------------------------------
+
+Added
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+#. `#235 <https://github.com/moremoban/moban/issues/235>`_: user defined
+   template types so that custom file extensions, template configurations can be
+   controlled by moban user
+#. `#232 <https://github.com/moremoban/moban/issues/232>`_: the package
+   dependencies have been fine tuning to lower versions, most of them are dated
+   back to 2017.
+
 0.4.0 - 20.02.2019
 --------------------------------------------------------------------------------